Lesson 10

+++ src/Controller/Api/v1/UserController.php
@@ -2,31 +2,45 @@
 
 namespace App\Controller\Api\v1;
 
+use App\DTO\UserDTO;
 use App\Entity\User;
+use App\Security\Voter\UserVoter;
 use App\Service\UserService;
+use JsonException;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 /** @Route("/api/v1/user") */
 class UserController
 {
-    /** @var UserService */
-    private $userService;
+    private UserService $userService;
 
-    public function __construct(UserService $userService)
+    private AuthorizationCheckerInterface $authorizationChecker;
+
+    public function __construct(UserService $userService, AuthorizationCheckerInterface $authorizationChecker)
     {
         $this->userService = $userService;
+        $this->authorizationChecker = $authorizationChecker;
     }
 
     /**
      * @Route("", methods={"POST"})
+     *
+     * @throws JsonException
      */
     public function saveUserAction(Request $request): Response
     {
-        $login = $request->request->get('login');
-        $userId = $this->userService->saveUser($login);
+        $userDTO = new UserDTO(
+            [
+                'login' => $request->request->get('login'),
+                'password' => $request->request->get('password'),
+                'roles' => $request->request->get('roles'),
+            ]
+        );
+        $userId = $this->userService->saveUser(new User(), $userDTO);
         [$data, $code] = $userId === null ?
             [['success' => false], 400] :
             [['success' => true, 'userId' => $userId], 200];
@@ -52,6 +66,10 @@ class UserController
      */
     public function deleteUserAction(int $id): Response
     {
+        $user = $this->userService->findUserById($id);
+        if (!$this->authorizationChecker->isGranted(UserVoter::DELETE, $user)) {
+            return new JsonResponse('Access denied', 403);
+        }
         $result = $this->userService->deleteUserById($id);
 
         return new JsonResponse(['success' => $result], $result ? 200 : 404);
@@ -62,9 +80,15 @@ class UserController
      */
     public function updateUserAction(Request $request): Response
     {
-        $userId = $request->query->get('userId');
-        $login = $request->query->get('login');
-        $result = $this->userService->updateUser($userId, $login);
+        $userId = $request->request->get('userId');
+        $userDTO = new UserDTO(
+            [
+                'login' => $request->request->get('login'),
+                'password' => $request->request->get('password'),
+                'roles' => $request->request->get('roles'),
+            ]
+        );
+        $result = $this->userService->updateUser($userId, $userDTO);
 
         return new JsonResponse(['success' => $result], $result ? 200 : 404);
     }
